Version 1.5.1: improved video undeletion support and more
* Fixed some fatals here and there under MediaWiki 1.24.1
* Instead of reusing certain core messages which mention "files", added our own versions of them
* Initial French (fr) i18n by me
* Moved log formatting to the post-MW 1.19 style; this obviously breaks old log entries, but it's of little concern since this extension hasn't been and still isn't production-ready
* Custom and extremely hacky class which overrides core Special:Undelete to allow undeleting videos without patching core
** The core patch is thus no longer required and has been deleted
* Fixed a strange video undeletion bug where undeleting a video which had only one history entry (the current one) would be restored as two separate entries (with the same timestamp etc.) -- not sure if this is caused by my Special:Undelete stuff or not
* Removed last wfMsg call

+++ VideoHooks.php
@@ -221,31 +221,33 @@ class VideoHooks {
 			// Delicious copypasta from Article.php, function doDeleteArticle()
 			// with some modifications
 			$archiveName = gmdate( 'YmdHis' ) . "!{$videoName}";
-			$dbw->begin();
-			$dbw->insertSelect(
-				'oldvideo',
-				'video',
-				array(
-					'ov_name' => 'video_name',
-					'ov_archive_name' => $dbw->addQuotes( $archiveName ),
-					'ov_url' => 'video_url',
-					'ov_type' => 'video_type',
-					'ov_user_id' => 'video_user_id',
-					'ov_user_name' => 'video_user_name',
-					'ov_timestamp' => 'video_timestamp'
-				),
-				$where,
-				__METHOD__
-			);
+			if ( !empty( $videoName ) ) {
+				$dbw->begin();
+				$dbw->insertSelect(
+					'oldvideo',
+					'video',
+					array(
+						'ov_name' => 'video_name',
+						'ov_archive_name' => $dbw->addQuotes( $archiveName ),
+						'ov_url' => 'video_url',
+						'ov_type' => 'video_type',
+						'ov_user_id' => 'video_user_id',
+						'ov_user_name' => 'video_user_name',
+						'ov_timestamp' => 'video_timestamp'
+					),
+					$where,
+					__METHOD__
+				);
 
-			// Now that it's safely backed up, delete it
-			$dbw->delete(
-				'video',
-				$where,
-				__METHOD__
-			);
+				// Now that it's safely backed up, delete it
+				$dbw->delete(
+					'video',
+					$where,
+					__METHOD__
+				);
 
-			$dbw->commit();
+				$dbw->commit();
+			}
 
 			// Purge caches
 			// None of these help, the video is still displayed on pages where
